• Added `checkmutes` command

+++ cogs/moderator.py
@@ -7,6 +7,21 @@ import datetime
 from discord.ext import commands, tasks
 from dateutil.relativedelta import relativedelta
 from utils import time
+from utils.paginator import CustomMenu
+
+class MutePageEntry:
+	def __init__(self, entry):
+		
+		self.name = entry['name']
+		self.time_left = entry['time_left']
+	
+	def __str__(self):
+		return f'**{self.name}** (`{self.time_left}`)'
+
+class MutePages(CustomMenu):
+	def __init__(self, entries, *, per_page=12, title="", color=None):
+		converted = [MutePageEntry(entry) for entry in entries]
+		super().__init__(converted, per_page=per_page, color=color, title=title)
 
 time_regex = re.compile("(?:(\d{1,5})(h|s|m|d))+?")
 time_dict = {"h":3600, "s":1, "m":60, "d":86400}
@@ -53,40 +68,42 @@ class Moderation(commands.Cog):
 		results = await self.db1.find().to_list(100000)
 		results2 = await self.db2.find().to_list(100000)
 		for result in results:
-			unmuteTime = result['mutedAt'] + relativedelta(seconds=result['muteDuration'])
+			if result['muteDuration'] != None:
+				unmuteTime = result['mutedAt'] + relativedelta(seconds=result['muteDuration'])
 
-			if currentTime >= unmuteTime:
-				guild = self.bot.get_guild(result['guildId'])
-				member = guild.get_member(result['_id'])
+				if currentTime >= unmuteTime:
+					guild = self.bot.get_guild(result['guildId'])
+					member = guild.get_member(result['_id'])
 
-				mute_role = guild.get_role(750465726069997658)
-				
-				if member != None:
-					if mute_role in member.roles:
-						await member.remove_roles(mute_role)
-						await member.send("You have been unmuted in `ViHill Corner`.")
+					mute_role = guild.get_role(750465726069997658)
 					
-					await self.db1.delete_one({"_id": member.id})
-				else:
-					await self.db1.delete_one({"_id": result['_id']})
+					if member != None:
+						if mute_role in member.roles:
+							await member.remove_roles(mute_role)
+							await member.send("You have been unmuted in `ViHill Corner`.")
+						
+						await self.db1.delete_one({"_id": member.id})
+					else:
+						await self.db1.delete_one({"_id": result['_id']})
 
 		for result2 in results2:
-			unmuteTime = result2['mutedAt'] + relativedelta(seconds=result2['muteDuration'])
+			if result2['muteDuration'] != None:
+				unmuteTime = result2['mutedAt'] + relativedelta(seconds=result2['muteDuration'])
 
-			if currentTime >= unmuteTime:
-				guild = self.bot.get_guild(result2['guildId'])
-				member = guild.get_member(result2['_id'])
+				if currentTime >= unmuteTime:
+					guild = self.bot.get_guild(result2['guildId'])
+					member = guild.get_member(result2['_id'])
 
-				mute_role = guild.get_role(750465726069997658)
-				
-				if member != None:
-					if mute_role in member.roles:
-						await member.remove_roles(mute_role)
-						await member.send("You have been unmuted in `ViHill Corner`.")
+					mute_role = guild.get_role(750465726069997658)
 					
-					await self.db2.delete_one({"_id": member.id})
-				else:
-					await self.db2.delete_one({"_id": result2['_id']})
+					if member != None:
+						if mute_role in member.roles:
+							await member.remove_roles(mute_role)
+							await member.send("You have been unmuted in `ViHill Corner`.")
+						
+						await self.db2.delete_one({"_id": member.id})
+					else:
+						await self.db2.delete_one({"_id": result2['_id']})
 
 	@commands.command()
 	@commands.has_role(754676705741766757)
@@ -239,85 +256,132 @@ class Moderation(commands.Cog):
 
 	@commands.command()
 	@commands.has_role(754676705741766757)
-	async def mute(self, ctx):
-		"""Mute someone permanently until you decide to unmute them... if you remember to."""
+	async def mute(self, ctx, member : discord.Member, *, muted_time : TimeConverter = None):
+		"""
+		Mutes a member for the specified time- time in 2d 10h 3m 2s format ex:
+		!mute @Someone 1d
+		"""
 
 		guild = self.bot.get_guild(750160850077089853)
 		log_channel = guild.get_channel(788377362739494943)
-		mem_list = []
-		muted = guild.get_role(750465726069997658)
 		staff = guild.get_role(754676705741766757)
-		
-		def check(m):
-			return m.author.id == ctx.author.id and m.channel.id == ctx.channel.id
 
-		await ctx.send("What member(s) do you wish to mute? To cancel type `cancel`")
-		try:
-			before_members = await self.bot.wait_for('message', timeout=180, check=check)
-			if before_members.content.lower() == "cancel":
-				await ctx.send("Canceled.")
-				return
-			else:
-				muted_members = before_members.mentions
+		if not staff in member.roles:
+			
+			def format_time(dt):
+				return time.human_timedelta(dt, accuracy = 3)
 
-		except asyncio.TimeoutError:
-			return
-		
-		else:
-			await ctx.send("What's the reason for the mute?")
+			def check(m):
+				return m.author.id == ctx.author.id and m.channel.id == ctx.channel.id
+			
+			await ctx.send("What's the reason?")
 			try:
-				before_reason = await self.bot.wait_for('message', timeout=360, check=check)
-				if before_reason.content.lower() == "cancel":
-					await ctx.send("Canceled.")
-					return
-				else:
-					mute_reason = before_reason.content
+				get_reason = await self.bot.wait_for('message', timeout=180, check=check)
+				reason_content = get_reason.content
 
 			except asyncio.TimeoutError:
-				return
+				return await ctx.reply("Reason is something you must give!")
 			
 			else:
-				for id in muted_members:
-					if not staff in id.roles:
-						post = {
-							'_id': id.id,
-							'muteDuration': None,
-							'mutedAt': datetime.datetime.now(),
-							'mutedBy': ctx.author.id,
-							'guildId': ctx.guild.id,
-						}
+				post = {
+						'_id': member.id,
+						'mutedAt': datetime.datetime.utcnow(),
+						'muteDuration': muted_time,
+						'mutedBy': ctx.author.id,
+						'guildId': ctx.guild.id,
+					}
 
+				try:
+					await self.db1.insert_one(post)
+				except:
+					await ctx.send("User is already muted!")
+					return
+
+				if muted_time != None:
+					muted_for = datetime.datetime.utcnow() + relativedelta(seconds = muted_time)
+					muted_for = format_time(muted_for)
+				else:
+					muted_for = "Eternity"
+
+				muted = guild.get_role(750465726069997658)
+				await member.add_roles(muted, reason=f'{ctx.author}: "{reason_content}"')
+				msg = "You have been muted in `ViHill Corner`"
+				em = discord.Embed(description=f"Time: `{muted_for}`\n**Reason: [{reason_content}]({ctx.message.jump_url})**", color=color.inviscolor)
+				try:
+					await member.send(msg, embed = em)
+				except discord.HTTPException:
+					pass
+
+
+				unban = discord.Embed(description= f'{member.mention} has been temporarily muted. \n\nTime: `{muted_for}`\n**Reason: [{reason_content}]({ctx.message.jump_url})**' , color=color.red)
+				
+				await ctx.send(embed=unban)
+
+				log = discord.Embed(color=color.reds, title="___Mute___", timestamp = ctx.message.created_at)
+				log.add_field(name="Member", value=f"`{member}`", inline=False)
+				log.add_field(name="Moderator", value=f"`{ctx.author}`", inline=False)
+				log.add_field(name="Time", value=f"`{muted_for}`", inline=False)
+				log.add_field(name="Reason", value=f"**[{reason_content}]({ctx.message.jump_url})**", inline=False)
+				await log_channel.send(embed=log)
+
+				if muted_time != None:
+					await asyncio.sleep(muted_time)
+					if muted in member.roles:
+						await member.remove_roles(muted)
+						await self.db1.delete_one({"_id": member.id})
 						try:
-							await self.db1.insert_one(post)
-						except:
-							await ctx.send("User is already muted!")
-							return
-						mute = discord.Embed(description=f'**Reason:** **[{mute_reason}]({ctx.message.jump_url}).**', color=color.inviscolor)
-						msg="You were muted in `ViHill Corner`."
-						a = id
-						mem_list.append(a)
-						mem_list_final = " | ".join(str(id) for id in mem_list)
-						try:
-							await id.send(msg, embed=mute)
+							await member.send("You have been unmuted in `ViHill Corner`.")
 						except discord.HTTPException:
 							pass
-						await id.add_roles(muted, reason=f"{ctx.author} ---> {mute_reason}")
+					else:
+						pass
 
-		ban = discord.Embed(description=f"The user(s) have been muted!\n**Reason:** **[{mute_reason}]({ctx.message.jump_url})**" , color=discord.Color.red())
+		else:
+			await ctx.send("You can't mute mods or take any moderator action against them.")
+			return
 
-		await ctx.send(embed=ban)
+	@commands.command(name='checkmute', aliases=['checkmutes'])
+	async def check_mutes(self, ctx, member: discord.Member = None):
+		"""Check to see if the member is muted if specified any, or in case no member is specified then see all the members that are muted if any."""
 
-		em = discord.Embed(color=color.reds, title="___MUTE___", timestamp = ctx.message.created_at)	
-		em.add_field(name="Moderator", value=f"`{ctx.author}`", inline=False)	
-		em.add_field(name="Action", value="`Used the mute command.`", inline=False)	
-		try:
-			em.add_field(name="Member(s)", value=f"`{mem_list_final}`", inline=False)	
-		except UnboundLocalError:
-			em.add_field(name="Member(s)", value="`Invalid Users!`", inline=False)		
-		em.add_field(name="Reason", value=f"**[{mute_reason}]({ctx.message.jump_url})**", inline=False)	
-		em.add_field(name="Channel", value=f"<#{ctx.channel.id}>", inline=False)
+		if member is None:
+			entries = []
+			results1 = await self.db1.find().to_list(100000)
+			results2 = await self.db2.find().to_list(100000)
+			results = results1+results2
+			if results == None:
+				return await ctx.reply("No members muted currently.")
+			for result in results:
+				if result['muteDuration'] != None:
+					_time = datetime.datetime.utcnow() + relativedelta(seconds = result['muteDuration'])
+					_time = f"Time Left: {time.human_timedelta(_time, suffix=False, brief=True)}"
+				else:
+					_time = "Time Left: Eternity"
+				username = self.bot.get_user(result['_id'])
+				_dict = {'username': username, 'time_left': _time}
+				entries.append(_dict)
+			m = MutePages(entries, per_page=5, title="Here's all the current muted members:", color=color.red)
+			await m.start(ctx)
+		
+		else:
+			result = await self.db1.find_one({'_id': member.id})
+			if result is None:
+				result = await self.db2.find_one({'_id': member.id})
+				if result is None:
+					return await ctx.reply("That member is not muted.")
 
-		await log_channel.send(embed=em)
+			if result['muteDuration'] != None:
+				_time = datetime.datetime.utcnow() + relativedelta(seconds = result['muteDuration'])
+				_time = f"Time Left: {time.human_timedelta(_time, suffix=False)}"
+			else:
+				_time = "Eternity"
+			
+			em = discord.Embed(color=color.red)
+			em.set_author(name=member, url=member.avatar_url, icon_url=member.avatar_url)
+			em.description = f"Time Left: `{_time}`"
+			em.set_footer(text=f"Requested by: {ctx.author}", icon_url=ctx.author.avatar_url)
+
+			await ctx.send(embed=em)			
 
 	@commands.command()
 	@commands.has_role(754676705741766757)
@@ -411,96 +475,10 @@ class Moderation(commands.Cog):
 
 		await log_channel.send(embed=em)
 
-	@commands.command()
-	@commands.has_role(754676705741766757)
-	async def tempmute(self, ctx, member : discord.Member, *, muted_time : TimeConverter = None):
-		"""
-		Mutes a member for the specified time- time in 2d 10h 3m 2s format ex:
-		!mute @Someone 1d
-		"""
-
-		guild = self.bot.get_guild(750160850077089853)
-		log_channel = guild.get_channel(788377362739494943)
-		staff = guild.get_role(754676705741766757)
-
-		if not staff in member.roles:
-			
-			def format_time(dt):
-				return time.human_timedelta(dt, accuracy = 3)
-
-			def check(m):
-				return m.author.id == ctx.author.id and m.channel.id == ctx.channel.id
-			
-			await ctx.send("What's the reason?")
-			try:
-				get_reason = await self.bot.wait_for('message', timeout=180, check=check)
-				reason_content = get_reason.content
-
-			except asyncio.TimeoutError:
-				return
-			
-			else:
-
-				if time is None:
-					await ctx.send("You need to specify time.")
-					return
-				else:
-					post = {
-							'_id': member.id,
-							'mutedAt': datetime.datetime.now(),
-							'muteDuration': muted_time,
-							'mutedBy': ctx.author.id,
-							'guildId': ctx.guild.id,
-						}
-
-					try:
-						await self.db1.insert_one(post)
-					except:
-						await ctx.send("User is already muted!")
-						return
-
-					muted_for = datetime.datetime.utcnow() + relativedelta(seconds = muted_time)
-
-					muted = guild.get_role(750465726069997658)
-					await member.add_roles(muted, reason=f"{ctx.author} ---> {reason_content}")
-					msg = ("You have been muted in `ViHill Corner`")
-					em = discord.Embed(description=f"Time: `{format_time(muted_for)}`\n**Reason: [{reason_content}]({ctx.message.jump_url})**", color=color.inviscolor)
-					try:
-						await member.send(msg, embed = em)
-					except discord.HTTPException:
-						pass
-
-
-					unban = discord.Embed(description= f'{member.mention} has been temporarily muted. \n\nTime: `{format_time(muted_for)}`\n**Reason: [{reason_content}]({ctx.message.jump_url})**' , color=color.red)
-					
-					await ctx.send(embed=unban)
-
-					log = discord.Embed(color=color.reds, title="___Mute___", timestamp = ctx.message.created_at)
-					log.add_field(name="Member", value=f"`{member}`", inline=False)
-					log.add_field(name="Moderator", value=f"`{ctx.author}`", inline=False)
-					log.add_field(name="Time", value=f"`{format_time(muted_for)}`", inline=False)
-					log.add_field(name="Reason", value=f"**[{reason_content}]({ctx.message.jump_url})**", inline=False)
-					await log_channel.send(embed=log)
-
-					await asyncio.sleep(muted_time)
-					if muted in member.roles:
-						await member.remove_roles(muted)
-						await self.db1.delete_one({"_id": member.id})
-						try:
-							await member.send("You have been unmuted in `ViHill Corner`.")
-						except discord.HTTPException:
-							pass
-					else:
-						pass
-
-		else:
-			await ctx.send("You can't mute mods or take any moderator action against them.")
-			return
-
 	@commands.command(name='purge', aliases=['clear'])
 	@commands.has_role(754676705741766757)
 	async def mod_purge(self, ctx, amount=0):
-		"""Delete the amount of messages from the chat"""
+		"""Delete the amount of messages from the chat."""
 
 		await ctx.message.delete()
 		await ctx.channel.purge(limit=amount)
